--- packages/types/tsconfig.cjs.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.cjs.json",
-  "compilerOptions": {
-    "outDir": "./lib"
-  },
-  "include": ["./src/**/*.ts"]
-}
